+++ src/Symfony/Component/PropertyInfo/PropertyInfoExtractor.php
@@ -20,7 +20,7 @@ use Symfony\Component\TypeInfo\Type;
  *
  * @final
  */
-class PropertyInfoExtractor implements PropertyInfoExtractorInterface, PropertyInitializableExtractorInterface
+class PropertyInfoExtractor implements PropertyInfoExtractorInterface, PropertyNameExtractorInterface, PropertyInitializableExtractorInterface
 {
     /**
      * @param iterable<mixed, PropertyListExtractorInterface>          $listExtractors
@@ -28,6 +28,7 @@ class PropertyInfoExtractor implements PropertyInfoExtractorInterface, PropertyI
      * @param iterable<mixed, PropertyDescriptionExtractorInterface>   $descriptionExtractors
      * @param iterable<mixed, PropertyAccessExtractorInterface>        $accessExtractors
      * @param iterable<mixed, PropertyInitializableExtractorInterface> $initializableExtractors
+     * @param iterable<mixed, PropertyNameExtractorInterface>          $propertyNameExtractors
      */
     public function __construct(
         private readonly iterable $listExtractors = [],
@@ -35,6 +36,7 @@ class PropertyInfoExtractor implements PropertyInfoExtractorInterface, PropertyI
         private readonly iterable $descriptionExtractors = [],
         private readonly iterable $accessExtractors = [],
         private readonly iterable $initializableExtractors = [],
+        private readonly iterable $propertyNameExtractors = [],
     ) {
     }
 
@@ -72,6 +74,11 @@ class PropertyInfoExtractor implements PropertyInfoExtractorInterface, PropertyI
     public function isWritable(string $class, string $property, array $context = []): ?bool
     {
         return $this->extract($this->accessExtractors, 'isWritable', [$class, $property, $context]);
+    }
+
+    public function getPropertyName(string $class, string $method, array $context = []): ?string
+    {
+        return $this->extract($this->propertyNameExtractors, 'getPropertyName', [$class, $method, $context]);
     }
 
     public function isInitializable(string $class, string $property, array $context = []): ?bool